requirements.txt
# Solo per: python3 check_cash.py (balance USDC via executor)
python-dotenv>=1.0.0
py-clob-client>=0.29.0
web3>=6.0.0
tenacity>=8.2.0
httpx>=0.27.0
